% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACMconcatenated.R
\name{ACMconcatenated}
\alias{ACMconcatenated}
\title{ACM Concatenated}
\usage{
ACMconcatenated(base, IndK, interactive = TRUE)
}
\arguments{
\item{base}{Data set}

\item{IndK}{Character with the name of the column that specifies the partition of the data set in k tables.}

\item{interactive}{If it is TRUE, the graph will be shown interactively. If FALSE, the graph is displayed flat. FALSE is the default.}
}
\value{
A Multiple Correspondence Analysis graph of the concatenated table.
}
\description{
Multiple correspondence analysis applied to a concatenated table.
}
\examples{
data(Datak10Contaminated)
ACMconcatenated(Datak10Contaminated,"GroupLetter", interactive = TRUE)
}
