% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACMpoint.R
\name{ACMpoint}
\alias{ACMpoint}
\title{ACM in one point}
\usage{
ACMpoint(base, IndK, PointTable, interactive = TRUE)
}
\arguments{
\item{base}{Data set}

\item{IndK}{Character with the name of the column that specifies the partition of the data set in k tables.}

\item{PointTable}{Table indicator. A character or number that is part of the \code{IndK} registers. This argument specifies the table to which the analysis will be performed.}

\item{interactive}{If it is TRUE, the graph will be shown interactively. If FALSE, the graph is displayed flat. FALSE is the default.}
}
\value{
A Multiple Correspondence Analysis graph of the table specified in \code{PointTable}.
}
\description{
Multiple correspondence analysis applied to a specific table.
}
\examples{
data(Datak10Contaminated)
ACMpoint(Datak10Contaminated,"GroupLetter", PointTable="j", interactive=TRUE)
}
