% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSq_variable.R
\name{ChiSq_variable}
\alias{ChiSq_variable}
\title{Chi squared variable from point table.}
\usage{
ChiSq_variable(base, IndK, PointTable, interactive = FALSE)
}
\arguments{
\item{base}{Data set}

\item{IndK}{Character with the name of the column that specifies the partition of the data set in k tables.}

\item{PointTable}{Table indicator. A character or number that is part of the \code{IndK} registers. This argument specifies the table to which the analysis will be performed.}
}
\value{
A table with Chi square distances between the column masses of the table specified in \code{PointTable} and the consolidated table. It has an indicator of significance with stars based on a scale.
}
\description{
Contains Chi square distance between the column masses of the table specified in \code{PointTable} and the consolidated table. It allows to identify which mode is responsible for the anomaly in the table in which it is located.
}
\examples{
data(Datak10Contaminated)
ChiSq_variable(Datak10Contaminated, "GroupLetter", PointTable="j")
}
