% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datak10Contaminated}
\alias{Datak10Contaminated}
\title{10 Tables Data Set}
\format{
A data frame:
\describe{
  \item{V1}{Contains 3 modes "High", "Medium", "Low".}
  \item{V2}{Contains 3 modes "High", "Medium", "Low".}
  \item{V3}{Contains 3 modes "High", "Medium", "Low".}
  \item{V4}{Contains 3 modes "High", "Medium", "Low".}
  \item{V5}{Contains 3 modes "High", "Medium", "Low".}
  \item{V6}{Contains 3 modes "High", "Medium", "Low".}
  \item{V7}{Contains 3 modes "High", "Medium", "Low".}
  \item{V8}{Contains 3 modes "High", "Medium", "Low".}
  \item{V9}{Contains 3 modes "High", "Medium", "Low".}
  \item{V10}{Contains 3 modes "High", "Medium", "Low".}
  \item{GroupLetter}{Letters from "a" to "j" identify the k tables.}
}
}
\usage{
Datak10Contaminated
}
\description{
Data from 10 tables with 10 categorical variables, the data from table 10 was generated with a different distribution from the others.
}
\keyword{datasets}
