% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datak10Contaminated}
\alias{Datak10Contaminated}
\title{10 Tables Data Set}
\format{
A data frame:
\describe{
  \item{V01}{Contains 3 modes "High", "Medium", "Low".}
  \item{V02}{Contains 3 modes "High", "Medium", "Low".}
  \item{V03}{Contains 3 modes "High", "Medium", "Low".}
  \item{V04}{Contains 3 modes "High", "Medium", "Low".}
  \item{V05}{Contains 3 modes "High", "Medium", "Low".}
  \item{V06}{Contains 3 modes "High", "Medium", "Low".}
  \item{V07}{Contains 3 modes "High", "Medium", "Low".}
  \item{V08}{Contains 3 modes "High", "Medium", "Low".}
  \item{V09}{Contains 3 modes "High", "Medium", "Low".}
  \item{V10}{Contains 3 modes "High", "Medium", "Low".}
  \item{GroupLetter}{Letters from "a" to "j" identify the k tables.}
}
}
\usage{
Datak10Contaminated
}
\description{
Data from 10 tables with 10 categorical variables, the data from table 10 was generated with a different distribution from the others.
}
\keyword{datasets}
